% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{consensus}
\alias{consensus}
\title{Consensus matrix computation}
\usage{
consensus(matrix, k)
}
\arguments{
\item{matrix}{a matrix containing co-association matrix}

\item{k}{number of clusters}
}
\description{
Computes consensus matrix given a co-association matrix
}
