% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{consmx}
\alias{consmx}
\title{Co-association matrix computation}
\usage{
consmx(dat)
}
\arguments{
\item{dat}{a matrix containing clustering solutions in columns}
}
\description{
Computes co-association matrix given cluster labels
res /= dat.n_cols;
}
